+++ src/resources/index.ts
@@ -28,6 +28,8 @@ export * from "./opportunity";
 export * from "./activity";
 export * from "./contract";
 export * from "./sales";
+export * from "./phase";
+export * from "./resource-list";
 export * from "./process";
 export * from "./resume";
 export * from "./attachment";